+++ changelog.md
@@ -10,6 +10,8 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 ## Fixed
 
+* `governance init` when genesis utxo had a script attached, then transaction fee was sometimes calculated incorrectly
+
 ## Added
 
 # v1.6.0